requirements.txt
bleak==3.0.0
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
meshtastic==2.7.8
packaging==24.2
protobuf==7.34.1
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreBluetooth==12.1
pyobjc-framework-libdispatch==12.1
pyotp==2.9.0
Pypubsub==4.0.7
pyserial==3.5
PyYAML==6.0.3
requests==2.32.5
tabulate==0.9.0
urllib3==2.6.3
